Updated code to fix errors and warnings on CI
- g++ 4.7 is no longer supported by TGUI 0.8, now using g++ 4.9
- valgrind was used by the script but not installed
- Fixed several errors for lines that VS2013 didn't support
- Fixed warnings in Any class about duplicate copy constructors
- Fixed some other minor warnings from VS2015

+++ src/TGUI/Widgets/SpinButton.cpp
@@ -53,7 +53,7 @@ namespace tgui
 
     /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-    SpinButton::Ptr SpinButton::create(float minimum, float maximum)
+    SpinButton::Ptr SpinButton::create(int minimum, int maximum)
     {
         auto spinButton = std::make_shared<SpinButton>();
